requirements.txt


Pillow
numpy
pytest
pytesseract
Flask-Cors
python-docx
markdown2
jinja2
python-pptx
